+++ app/providers.js
@@ -1,7 +1,7 @@
 'use client';
 
-import SyncContextProvider from '@/components/Sync/SyncContextProvider';
 import ScribexContextProvider from '@/components/context/ScribexContext';
+import SyncContextProvider from '@/components/context/SyncContext';
 import ProjectContextProvider from '../renderer/src/components/context/ProjectContext';
 import ReferenceContextProvider from '../renderer/src/components/context/ReferenceContext';
 import AuthenticationContextProvider from '../renderer/src/components/Login/AuthenticationContextProvider';
